move lua 5.3 compat things
