+++ data/lua/lua_5_3_compat.lua
@@ -1,0 +1,12 @@
+function bit.rshift(a, b)
+  return a >> b
+end
+function bit.lshift(a, b)
+  return a << b
+end
+function bit.bor(a, b)
+  return a | b
+end
+function bit.band(a, b)
+  return a & b
+end
